major restructuring into classes. Example scripts are converted to notebooks.
