--- e3dsecs/__init__.py
@@ -1,1 +1,0 @@
-from . import secs3dfrom . import gemini_toolsfrom . import visualizationfrom . import diagnosticsfrom . import uncertaintyfrom . import coordinates
