+++ paper/script_version/01_3dgrid_figure.py
