+++ paper/script_version/02_lompefit_example.py
