+++ paper/script_version/03_reconstruction_minimal.py
