+++ paper/script_version/05_additional_figures.py
